+++ tests/test_bracket/test_dhs.py
@@ -11,7 +11,6 @@ from autode.bracket.dhs import (
     DHS,
     DHSGS,
     DistanceConstrainedOptimiser,
-    TruncatedTaylor,
     DHSImagePair,
     ImageSide,
     OptimiserStepError,
@@ -22,33 +21,6 @@ from ..testutils import requires_working_xtb_install, work_in_zipped_dir
 
 here = os.path.dirname(os.path.abspath(__file__))
 datazip = os.path.join(here, "data", "geometries.zip")
-
-
-@requires_working_xtb_install
-@work_in_tmp_dir()
-def test_truncated_taylor_surface():
-    mol = Molecule(smiles="CCO")
-    mol.calc_hessian(method=XTB())
-    coords = CartesianCoordinates(mol.coordinates)
-    coords.update_g_from_cart_g(mol.gradient)
-    coords.update_h_from_cart_h(mol.hessian)
-    coords.make_hessian_positive_definite()
-
-    # for positive definite hessian, minimum of taylor surface would
-    # be a simple Newton step
-    minim = coords - (np.linalg.inv(coords.h) @ coords.g)
-
-    # minimizing surface should give the same result
-    surface = TruncatedTaylor(coords, coords.g, coords.h)
-    res = minimize(
-        method="CG",
-        fun=surface.value,
-        x0=np.array(coords),
-        jac=surface.gradient,
-    )
-
-    assert res.success
-    assert np.allclose(res.x, minim, rtol=1e-4)
 
 
 @requires_working_xtb_install
@@ -199,7 +171,8 @@ def test_dhs_single_step():
         initial_species=reactant,
         final_species=product,
         maxiter=200,
-        step_size=step_size,
+        large_step=0.2,
+        switch_thresh=1.5,
         dist_tol=1.0,
     )
 
@@ -213,13 +186,14 @@ def test_dhs_single_step():
     old_dist = imgpair.dist
     assert imgpair.left_coords.e > imgpair.right_coords.e
 
+    assert dhs.imgpair.dist > 1.5
     # take a single step
     dhs._step()
     # step should be on lower energy image
     assert len(imgpair._left_history) == 1
     assert len(imgpair._right_history) == 2
     new_dist = imgpair.dist
-    # image should move exactly by step_size
+    # image should move exactly by large step
     assert np.isclose(old_dist - new_dist, step_size)
 
 
@@ -235,7 +209,8 @@ def test_dhs_gs_single_step(caplog):
         initial_species=reactant,
         final_species=product,
         maxiter=200,
-        step_size=step_size,
+        large_step=step_size,
+        switch_thresh=1.5,
         dist_tol=1.0,
         gs_mix=0.5,
     )
@@ -243,6 +218,7 @@ def test_dhs_gs_single_step(caplog):
     dhs_gs.imgpair.set_method_and_n_cores(method=XTB(), n_cores=1)
     dhs_gs._method = XTB()
     dhs_gs._initialise_run()
+    assert dhs_gs.imgpair.dist > 1.5
     # take one step
     with caplog.at_level("INFO"):
         dhs_gs._step()
@@ -278,8 +254,8 @@ def test_dhs_diels_alder():
     dhs = DHS(
         initial_species=reactant,
         final_species=product,
+        small_step=0.2,
         maxiter=200,
-        step_size=0.2,
         dist_tol=set_dist_tol,
         gtol=1.0e-3,
     )
@@ -341,7 +317,8 @@ def test_dhs_jumping_over_barrier(caplog):
         initial_species=reactant,
         final_species=product,
         maxiter=200,
-        step_size=0.6,
+        large_step=0.5,
+        small_step=0.5,
         dist_tol=0.3,  # smaller dist_tol also to make one side jump
         gtol=5.0e-4,
         barrier_check=True,
@@ -368,7 +345,8 @@ def test_dhs_stops_if_microiter_exceeded(caplog):
         initial_species=reactant,
         final_species=product,
         maxiter=10,
-        step_size=0.2,
+        large_step=0.2,
+        small_step=0.1,
         dist_tol=1.0,
         gtol=5.0e-4,
         barrier_check=True,